it now fully supports creating, reading, updating, deleting and searching by id. Data shows in UI and updates after each action.

+++ src/main/java/com/gempel/app/rest/Controller/ApiControllers.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class ApiControllers {
@@ -24,6 +25,11 @@ public class ApiControllers {
         return userRepo.findAll();
     }
 
+    @PostMapping(value = "/search/{id}")
+    public Optional<User> searchUser(@PathVariable long id) {
+        return userRepo.findById(id);
+    }
+
     @PostMapping(value = "/save")
     public String saveUser(@RequestBody User user) {
         userRepo.save(user);
@@ -32,20 +38,30 @@ public class ApiControllers {
 
     @PutMapping(value = "/update/{id}")
     public String updateUser(@PathVariable long id, @RequestBody User user) {
-        User updatedUser = userRepo.findById(id).get();
-        updatedUser.setFirstName(user.getFirstName());
-        updatedUser.setLastName(user.getLastName());
-        updatedUser.setAge(user.getAge());
-        updatedUser.setOccupation(user.getOccupation());
-        userRepo.save(updatedUser);
-        return "Updated...";
+        Optional<User> optional = userRepo.findById(id);
+        if (optional.isPresent()) {
+            User updatedUser = optional.get();
+            updatedUser.setFirstName(user.getFirstName());
+            updatedUser.setLastName(user.getLastName());
+            updatedUser.setAge(user.getAge());
+            updatedUser.setOccupation(user.getOccupation());
+            userRepo.save(updatedUser);
+            return "Updated...";
+        } else {
+            return "No such user found!";
+        }
     }
 
     @DeleteMapping(value = "/delete/{id}")
     public String deleteUser(@PathVariable long id) {
-        User deleteUser = userRepo.findById(id).get();
-        userRepo.delete(deleteUser);
-        return "Delete user with the id: " + id;
+        Optional<User> optional = userRepo.findById(id);
+        if (optional.isPresent()) {
+            User deleteUser = optional.get();
+            userRepo.delete(deleteUser);
+            return "Delete user with the id: " + id;
+        } else {
+            return "No such user found!";
+        }
     }
 
 }
